+++ src/hooks/useTodoInput.ts
@@ -1,0 +1,11 @@
+import { useRef } from 'react';
+
+export const useTodoInput = () => {
+  const inputRef = useRef<HTMLInputElement>(null);
+
+  const onFocus = () => {
+    inputRef.current?.focus();
+  };
+
+  return { inputRef, onFocus };
+};
